refactor(membership): MembershipEventConsumer를 user-service로 이동 - user-service에 MembershipEventConsumer + MembershipUpdateRequestEvent 추가 - user-service에 Kafka consumer 설정 추가 - 모놀리스에서 MembershipEventConsumer 및 테스트 삭제
