+++ user-service/src/main/kotlin/com/hoppingmall/user/consumer/MembershipUpdateRequestEvent.kt
@@ -1,0 +1,11 @@
+package com.hoppingmall.user.consumer
+
+import java.math.BigDecimal
+
+data class MembershipUpdateRequestEvent(
+    val eventId: String,
+    val userId: Long,
+    val orderId: Long,
+    val paymentId: Long,
+    val amount: BigDecimal
+)
